RAD-188 Hide scheduledDate by default
added scheduledDate to radiologyOrderCreationSegment in previous commit but it
the input element was visible by default and on page load the js was executed
which set it to hidden since default urgency="ROUTINE". this lead to the
element being visible for a short amount of time if page load is slow.

* hide scheduledDate input by default
* click scheduledDate if urgency ON_SCHEDULED_DATE is selected this opens date
widget and makes it smoother for the user