RAD-107 Enforce valid DICOM UID for org root
the org root UID (DICOM root UID) is entered as a global
property defaults to 2.25 but can be changed by the admin in the settings page.
however validation is missing on entry meaning if the admin configures a wrong
uid it will fail during creation of radiology orders. so the cause and the
event of failiure are separated.

* added RegexDatatypeHandler which was missing for the custom datatype
* configured the dicomUIDOrgRoot as RegexValidatedTextDatatype with a regex
ensuring proper DICOM UID is entered by the admin
* update test datasets GP of dicomUIDOrgRoot