Remove radiology_order.order_id autoIncrement
the radiology_order.order_id primary key was set to autoIncrement which makes
it possible to insert a new radiology_order using

"insert into radiology_order values (NULL);"

if there is a test_order matching the current auto_increment value of the
radiology_order. This should not be possible as this can cause a
radiology_order being attached to a test_order which is in fact not a
radiology_order.

Not-null constraint on radiology_order.order_id and Foreign key constraint to test_order.order_id
together ensure integrity of radiology_order data.